chore: rename package to make functions self explanatory
