+++ internal/cmd/manager/instance/run/cmd.go
@@ -38,8 +38,8 @@ import (
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/cmd/manager/instance/run/lifecycle"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/management/controller"
+	"github.com/cloudnative-pg/cloudnative-pg/internal/management/controller/externalservers"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/management/controller/roles"
-	"github.com/cloudnative-pg/cloudnative-pg/internal/management/controller/servers"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/management/controller/slots/runner"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/management/controller/tablespaces"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/management/istio"
@@ -261,7 +261,7 @@ func runSubCommand(ctx context.Context, instance *postgres.Instance) error {
 	}
 
 	setupLog.Info("starting external server manager")
-	if err := servers.NewExternalServersReconciler(instance, mgr.GetClient()).
+	if err := externalservers.NewReconciler(instance, mgr.GetClient()).
 		SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create external servers reconciler")
 		return err
